95 wiki poll service
Actions:

+ create new WikiPollService trait and add to app state
+ impl WikiPollService for PolisClient
+ refactor types and usage

Motivation:

+ allow usage of polis client from application state
+ more efficient than re-instantiating client in routes
+ allow easier mocking of polis functionality in tests
